requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-ajax-selects==2.1.0
django-dropbox-storage==0.4.1
django-environ==0.4.5
django-heroku==0.3.1
django-import-export==2.5.0
django-qr-code==2.1.0
django-storages==1.11.1
dropbox==11.5.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
img2pdf==0.4.0
jdcal==1.4.1
kiwisolver==1.3.1
lml==0.1.0
lxml==4.6.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pikepdf==2.5.2
Pillow==8.1.0
ply==3.11
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
segno==1.3.1
six==1.15.0
sqlparse==0.4.1
stone==3.2.1
tablib==3.0.0
urllib3==1.26.4
virtualenv==20.4.3
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
